+++ src/highlightElement.js
@@ -11,8 +11,6 @@ debug.flag = false;
 
 const minWidth = 68;
 const minHeight = 27;
-const offset = 6;
-const borderWidth = 2;
 
 import {
   HIGHLIGHT_ID
@@ -30,31 +28,45 @@ styleHighlightTemplate.textContent = `
   margin: 0;
   padding: 0;
   position: absolute;
-  border-radius: 3px;
-  border: 4px solid light-dark($buttonBackgroundColor, $buttonBackgroundDarkColor);
+  border-radius: $highlightOffsetpx;
+  border: $shadowBorderWidthpx solid light-dark($menuBackgroundColor, $menuBackgroundDarkColor);
   box-sizing: border-box;
   pointer-events:none;
+  z-index: $zHighlight;
+}
+
+#${HIGHLIGHT_ID}.hasInfoBottom,
+#${HIGHLIGHT_ID} .overlay-border.hasInfoBottom {
+  border-radius: $highlightOffsetpx $highlightOffsetpx $highlightOffsetpx 0;
+}
+
+#${HIGHLIGHT_ID}.hasInfoTop,
+#${HIGHLIGHT_ID} .overlay-border.hasInfoTop {
+  border-radius: 0 $highlightOffsetpx $highlightOffsetpx $highlightOffsetpx;
 }
 
 #${HIGHLIGHT_ID} .overlay-border {
   margin: 0;
   padding: 0;
   position: relative;
-  top: -2px;
-  left: -2px;
-  border-radius: 3px 3px 3px 3px;
-  border: 2px solid light-dark($focusBorderColor, $focusBorderDarkColor);
+  border-radius: $highlightOffsetpx;
+  border: $overlayBorderWidthpx $highlightBorderStyle light-dark($focusBorderColor, $focusBorderDarkColor);
   z-index: $zHighlight;
   box-sizing: border-box;
   pointer-events:none;
+  background: transparent;
 }
+
 
 @keyframes fadeIn {
   0% { opacity: 0; }
   100% { opacity: 1; }
 }
 
-#${HIGHLIGHT_ID} .overlay-border.skip-to-hidden {
+#hidden-elem-msg {
+  position: absolute;
+  margin: 0;
+  padding: .25em;
   background-color: light-dark($hiddenHeadingBackgroundColor, $hiddenHeadingBackgroundDarkColor);
   color: light-dark($hiddenHeadingColor, $hiddenHeadingDarkColor);
   font-family: $fontFamily;
@@ -62,26 +74,18 @@ styleHighlightTemplate.textContent = `
   font-style: italic;
   font-weight: bold;
   text-align: center;
-  padding: .25em;
   animation: fadeIn 1.5s;
-}
-
-#${HIGHLIGHT_ID} .overlay-border.hasInfoBottom {
-  border-radius: 3px 3px 3px 0;
-}
-
-#${HIGHLIGHT_ID} .overlay-border.hasInfoTop {
-  border-radius: 0 3px 3px 3px;
+  z-index: $zHighlight;
 }
 
 #${HIGHLIGHT_ID} .overlay-info {
+  margin: 0;
+  padding: 2px;
   position: relative;
   text-align: left;
-  left: -2px;
-  padding: 1px 4px;
   font-size: $fontSize;
   font-family: $fontFamily;
-  border: 2px solid light-dark($focusBorderColor, $focusBorderDarkColor);
+  border: $infoBorderWidthpx solid light-dark($menuBackgroundColor, $menuBackgroundDarkColor);
   background-color: light-dark($menuBackgroundColor, $menuBackgroundDarkColor);
   color: light-dark($menuTextColor, $menuTextDarkColor);
   z-index: $zHighlight;
@@ -91,11 +95,11 @@ styleHighlightTemplate.textContent = `
 }
 
 #${HIGHLIGHT_ID} .overlay-info.hasInfoTop {
-  border-radius: 3px 3px 0 0;
+  border-radius: $highlightOffsetpx $highlightOffsetpx 0 0;
 }
 
 #${HIGHLIGHT_ID} .overlay-info.hasInfoBottom {
-  border-radius: 0 0 3px 3px;
+  border-radius: 0 0 $highlightOffsetpx $highlightOffsetpx;
 }
 
 @media (forced-colors: active) {
@@ -148,6 +152,17 @@ export default class HighlightElement extends HTMLElement {
     this.infoElem.className = 'overlay-info';
     this.overlayElem.appendChild(this.infoElem);
 
+    this.hiddenElem = document.createElement('div');
+    this.hiddenElem.id = 'hidden-elem-msg';
+    this.shadowRoot.appendChild(this.hiddenElem);
+    this.hiddenElem.style.display = 'none';
+
+    this.borderWidth    = 0;
+    this.borderContrast = 0;
+    this.offset         = 0;
+
+    this.msgHeadingIsHidden = '';
+
     this.configureStyle();
 
   }
@@ -160,7 +175,6 @@ export default class HighlightElement extends HTMLElement {
    *   @param {Object} config : color and font information
    */
 
-
   configureStyle(config={}) {
 
     function updateOption(style, option, configOption, defaultOption) {
@@ -172,6 +186,21 @@ export default class HighlightElement extends HTMLElement {
       }
     }
 
+    // Get i18n Messages
+
+    this.msgHeadingIsHidden = typeof config.msgHeadingIsHidden === 'string' ?
+                            config.msgHeadingIsHidden :
+                            'Heading is hidden';
+
+    this.msgRegionIsHidden = typeof config.msgRegionIsHidden === 'string' ?
+                            config.msgRegionIsHidden :
+                            'Region is hidden';
+
+    this.msgElementIsHidden = typeof config.msgElementIsHidden === 'string' ?
+                            config.msgElemenIsHidden :
+                            'Element is hidden';
+
+
     // make a copy of the template
     let style = styleHighlightTemplate.textContent.slice(0);
 
@@ -179,11 +208,6 @@ export default class HighlightElement extends HTMLElement {
                          '$fontFamily',
                          config.fontFamily,
                          defaultStyleOptions.fontFamily);
-
-    style = updateOption(style,
-                         '$fontSize',
-                         config.fontSize,
-                         defaultStyleOptions.fontSize);
 
     style = updateOption(style,
                          '$buttonBackgroundColor',
@@ -250,6 +274,76 @@ export default class HighlightElement extends HTMLElement {
                          config.zHighlight,
                          defaultStyleOptions.zHighlight);
 
+    style = updateOption(style,
+                         '$highlightBorderStyle',
+                         config.highlightBorderStyle,
+                         defaultStyleOptions.highlightBorderStyle);
+
+    const highlightBorderSize =  config.highlightBorderSize ?
+                                 config.highlightBorderSize :
+                                 defaultStyleOptions.highlightBorderSize;
+
+    switch (highlightBorderSize) {
+      case 'small':
+        this.borderWidth = 2;
+        this.borderContrast = 1;
+        this.offset = 4;
+        this.fontSize = '12pt';
+        break;
+
+      case 'medium':
+        this.borderWidth = 3;
+        this.borderContrast = 2;
+        this.offset = 4;
+        this.fontSize = '13pt';
+        break;
+
+      case 'large':
+        this.borderWidth = 4;
+        this.borderContrast = 3;
+        this.offset = 6;
+         this.fontSize = '14pt';
+       break;
+
+      case 'x-large':
+        this.borderWidth = 6;
+        this.borderContrast = 3;
+        this.offset = 8;
+        this.fontSize = '16pt';
+        break;
+
+      default:
+        this.borderWidth = 2;
+        this.borderContrast = 1;
+        this.offset = 4;
+        this.fontSize = '12pt';
+        break;
+    }
+
+    style = updateOption(style,
+                         '$fontSize',
+                         this.fontSize,
+                         defaultStyleOptions.fontSize);
+
+    style = updateOption(style,
+                         '$highlightOffset',
+                         this.offset,
+                         this.offset);
+
+    style = updateOption(style,
+                         '$overlayBorderWidth',
+                         this.borderWidth,
+                         this.borderWidth);
+
+    style = updateOption(style,
+                         '$shadowBorderWidth',
+                         this.borderWidth + 2 * this.borderContrast,
+                         this.borderWidth + 2 * this.borderContrast);
+
+    style = updateOption(style,
+                         '$infoBorderWidth',
+                         this.borderWidth,
+                         this.borderWidth);
 
     let styleNode = this.shadowRoot.querySelector('style');
 
@@ -264,18 +358,216 @@ export default class HighlightElement extends HTMLElement {
   }
 
   /*
+   *   @method highlight
+   *
+   *   @desc  Highlights the element on the page when highlighting
+   *          is enabled (NOTE: Highlight is enabled by default)
+   *
+   *   @param {Object}  elem            : DOM node of element to highlight
+   *   @param {String}  highlightTarget : value of highlight target
+   *   @param {String}  info            : Information about target
+   *   @param {Boolean} force           : If true override isRduced
+   */
+
+  highlight(elem, highlightTarget, info='', force=false) {
+    let scrollElement;
+    const mediaQuery = window.matchMedia(`(prefers-reduced-motion: reduce)`);
+    const isReduced = !mediaQuery || mediaQuery.matches;
+
+    if (elem && highlightTarget) {
+
+      const rect = elem.getBoundingClientRect();
+
+      // If target element is hidden create a visible element
+      debug.flag && debug.log(`[    info]: ${info}`);
+      debug.flag && debug.log(`[    rect]: Left: ${rect.left} Top: ${rect.top} Width: ${rect.width} height: ${rect.height}`);
+      debug.flag && debug.log(`[isHidden]: ${this.isElementHidden(elem)}`);
+
+      if (this.isElementHidden(elem)) {
+        // If element is hidden make hidden element message visible
+        // and use for highlighing
+        this.hiddenElem.textContent = this.getHiddenMessage(elem);
+        this.hiddenElem.style.display = 'block';
+
+        const left = rect.left > 0 ? rect.left + window.scrollX : this.offset;
+        const top  = rect.top > 0 ? rect.top + window.scrollY : this.offset;
+
+        this.hiddenElem.style.left = left + 'px';
+        this.hiddenElem.style.top = top + 'px';
+        scrollElement = this.updateHighlightElement(this.hiddenElem,
+                                                    info,
+                                                    0,
+                                                    this.borderWidth,
+                                                    this.borderContrast);
+      }
+      else {
+        this.hiddenElem.style.display = 'none';
+        scrollElement = this.updateHighlightElement(elem,
+                                                    info,
+                                                    this.offset,
+                                                    this.borderWidth,
+                                                    this.borderContrast);
+      }
+
+      if (this.isElementInHeightLarge(elem)) {
+        if (!this.isElementStartInViewport(elem) && (!isReduced || force)) {
+          scrollElement.scrollIntoView({ behavior: highlightTarget, block: 'start', inline: 'nearest' });
+        }
+      }
+      else {
+        if (!this.isElementInViewport(elem)  && (!isReduced || force)) {
+          scrollElement.scrollIntoView({ behavior: highlightTarget, block: 'center', inline: 'nearest' });
+        }
+      }
+    }
+  }
+
+  /*
+   *  @method  updateHighlightElement
+   *
+   *  @desc  Create an overlay element and set its position on the page.
+   *
+   *  @param  {Object}  elem          -  DOM element node to highlight
+   *  @param  {String}  info          -  Description of the element
+   *  @param  {Number}  offset        -  Number of pixels for offset
+   *  @param  {Number}  borderWidth   -  Number of pixels for border width
+   *  @param  {Number}  borderContrast  -  Number of pixels to provide border contrast
+   *
+   */
+
+   updateHighlightElement (elem, info, offset, borderWidth, borderContrast) {
+
+    const adjRect = this.getAdjustedRect(elem, offset, borderWidth, borderContrast);
+
+    const borderElemOffset = -1 * (this.borderWidth + this.borderContrast);
+
+    this.overlayElem.style.left   = adjRect.left   + 'px';
+    this.overlayElem.style.top    = adjRect.top    + 'px';
+    this.borderElem.style.left    = borderElemOffset + 'px';
+    this.borderElem.style.top     = borderElemOffset + 'px';
+
+    this.overlayElem.style.width  = adjRect.width  + 'px';
+    this.overlayElem.style.height = adjRect.height + 'px';
+    this.borderElem.style.width   = (adjRect.width - (2 * borderContrast)) + 'px';
+    this.borderElem.style.height  = (adjRect.height - (2 * borderContrast)) + 'px';
+
+    this.overlayElem.style.display = 'block';
+
+    if (info) {
+
+      this.infoElem.style.display = 'inline-block';
+      this.infoElem.textContent   = info;
+
+      const infoElemOffsetLeft = -1 * (borderWidth + 2 * borderContrast);
+      this.infoElem.style.left = infoElemOffsetLeft + 'px';
+
+      const infoElemRect    = this.infoElem.getBoundingClientRect();
+
+      // Is info displayed above or below the highlighted element
+      if (adjRect.top >= infoElemRect.height) {
+        // Info is displayed above the highlighted element (e.g. most of the time)
+        this.overlayElem.classList.remove('hasInfoBottom');
+        this.borderElem.classList.remove('hasInfoBottom');
+        this.infoElem.classList.remove('hasInfoBottom');
+        this.overlayElem.classList.add('hasInfoTop');
+        this.borderElem.classList.add('hasInfoTop');
+        this.infoElem.classList.add('hasInfoTop');
+        this.infoElem.style.top =  (-1 * (adjRect.height +
+                                         infoElemRect.height +
+                                         borderWidth))  + 'px';
+      }
+      else {
+        // Info is displayed below the highlighted element when it is at the top of
+        // the window
+
+        const infoElemOffsetTop  = -1 * (borderWidth + borderContrast);
+
+        this.overlayElem.classList.remove('hasInfoTop');
+        this.borderElem.classList.remove('hasInfoTop');
+        this.infoElem.classList.remove('hasInfoTop');
+        this.overlayElem.classList.add('hasInfoBottom');
+        this.borderElem.classList.add('hasInfoBottom');
+        this.infoElem.classList.add('hasInfoBottom');
+        this.infoElem.style.top  = infoElemOffsetTop + 'px';
+      }
+      return this.infoElem;
+    }
+    else {
+      this.overlayElem.classList.remove('hasInfoTop');
+      this.overlayElem.classList.remove('hasInfoBottom');
+      this.borderElem.classList.remove('hasInfoTop');
+      this.borderElem.classList.remove('hasInfoBottom');
+      this.infoElem.style.display = 'none';
+      return this.overlayElem;
+    }
+  }
+
+
+  /*
+   *   @method getAdjustedRect
+   *
+   *   @desc  Returns a object with dimensions adjusted for highlighting element
+   *
+   *  @param  {Object}  elem            -  DOM node of element to be highlighted
+   *  @param  {Number}  offset          -  Number of pixels for offset
+   *  @param  {Number}  borderWidth     -  Number of pixels for border width
+   *  @param  {Number}  borderContrast  -  Number of pixels to provide border contrast
+   *
+   *   @returns see @desc
+   */
+
+   getAdjustedRect(elem, offset, borderWidth, borderContrast) {
+
+    const rect  = elem.getBoundingClientRect();
+
+    const adjRect = {
+      left: 0,
+      top: 0,
+      width: 0,
+      height: 0
+    };
+
+    const offsetBorder = offset + borderWidth + 2 * borderContrast;
+
+    adjRect.left    = rect.left > offset ?
+                      Math.round(rect.left + (-1 * offsetBorder) + window.scrollX) :
+                      Math.round(rect.left + window.scrollX);
+
+    adjRect.width   = rect.left > offset ?
+                      Math.max(rect.width  + (2 * offsetBorder), minWidth) :
+                      Math.max(rect.width, minWidth);
+
+
+    adjRect.top     = rect.top > offset ?
+                      Math.round(rect.top  + (-1 * offsetBorder) + window.scrollY) :
+                      Math.round(rect.top + window.scrollY);
+
+    adjRect.height  = rect.top > offset ?
+                      Math.max(rect.height + (2 * offsetBorder), minHeight) :
+                      Math.max(rect.height, minHeight);
+
+    if ((adjRect.top < 0) || (adjRect.left < 0)) {
+    // Element is near top or left side of screen
+      adjRect.left = this.offset;
+      adjRect.top = this.offset;
+    }
+
+    return adjRect;
+  }
+
+  /*
    *   @method isElementInViewport
    *
    *   @desc  Returns true if element is already visible in view port,
    *          otheriwse false
    *
-   *   @param {Object} element : DOM node of element to highlight
+   *   @param {Object} elem : DOM node of element to highlight
    *
    *   @returns see @desc
    */
 
-  isElementInViewport(element) {
-    const rect = element.getBoundingClientRect();
+  isElementInViewport(elem) {
+    const rect = elem.getBoundingClientRect();
     return (
       rect.top >= window.screenY &&
       rect.left >= window.screenX &&
@@ -285,20 +577,19 @@ export default class HighlightElement extends HTMLElement {
                      (window.screenX + document.documentElement.clientWidth)));
   }
 
-
   /*
    *   @method isElementStartInViewport
    *
    *   @desc  Returns true if start of the element is already visible in view port,
    *          otherwise false
    *
-   *   @param {Object} element : DOM node of element to highlight
+   *   @param {Object} elem : DOM node of element to highlight
    *
    *   @returns see @desc
    */
 
-  isElementStartInViewport(element) {
-    const rect = element.getBoundingClientRect();
+  isElementStartInViewport(elem) {
+    const rect = elem.getBoundingClientRect();
     return (
         rect.top >= window.screenY &&
         rect.top <= ((window.screenY + window.innerHeight) ||
@@ -315,47 +606,57 @@ export default class HighlightElement extends HTMLElement {
    *   @desc  Returns true if element client height is larger than clientHeight,
    *          otheriwse false
    *
-   *   @param {Object} element : DOM node of element to highlight
+   *   @param {Object} elem : DOM node of element to highlight
    *
    *   @returns see @desc
    */
 
-  isElementInHeightLarge(element) {
-    var rect = element.getBoundingClientRect();
+  isElementInHeightLarge(elem) {
+    var rect = elem.getBoundingClientRect();
     return (1.2 * rect.height) > (window.innerHeight || document.documentElement.clientHeight);
   }
 
   /*
-   *   @method highlight
+   *   @method isElementHidden
    *
-   *   @desc  Highlights the element with the id on a page when highlighting
-   *          is enabled (NOTE: Highlight is enabled by default)
+   *   @desc  Returns true if the element is hidden on the
+   *          graphical rendering
    *
-   *   @param {Object}  elem            : DOM node of element to highlight
-   *   @param {String}  highlightTarget : value of highlight target
-   *   @param {String}  info            : Information about target
-   *   @param {Boolean} force           : If true override isRduced
+   *   @param  {Object}  elem   : DOM node
+   *
+   *   @returns see @desc
    */
+  isElementHidden(elem) {
+    const rect = elem.getBoundingClientRect();
+    return (rect.height < 3) ||
+           (rect.width  < 3) ||
+           ((rect.left + rect.width)  < (rect.width / 2)) ||
+           ((rect.top  + rect.height) < (rect.height / 2));
+  }
 
-  highlight(elem, highlightTarget, info='', force=false) {
-    const mediaQuery = window.matchMedia(`(prefers-reduced-motion: reduce)`);
-    const isReduced = !mediaQuery || mediaQuery.matches;
+  /*
+   *   @method getHiddenMessage
+   *
+   *   @desc  Returns string describing the hidden element
+   *
+   *   @param  {Object}  elem   : DOM node
+   *
+   *   @returns see @desc
+   */
+  getHiddenMessage(elem) {
+    if (elem.hasAttribute('data-skip-to-info')) {
+      const info = elem.getAttribute('data-skip-to-info');
 
-    if (elem && highlightTarget) {
-
-      const scrollElement = this.updateHighlightElement(elem, info);
-
-      if (this.isElementInHeightLarge(elem)) {
-        if (!this.isElementStartInViewport(elem) && (!isReduced || force)) {
-          scrollElement.scrollIntoView({ behavior: highlightTarget, block: 'start', inline: 'nearest' });
-        }
+      if (info.includes('heading')) {
+        return this.msgHeadingIsHidden;
       }
-      else {
-        if (!this.isElementInViewport(elem)  && (!isReduced || force)) {
-          scrollElement.scrollIntoView({ behavior: highlightTarget, block: 'center', inline: 'nearest' });
-        }
+
+      if (info.includes('landmark')) {
+        return this.msgRegionIsHidden;
       }
     }
+
+    return this.msgElementIsHidden;
   }
 
   /*
@@ -364,135 +665,9 @@ export default class HighlightElement extends HTMLElement {
    *   @desc  Hides the highlight element on the page
    */
   removeHighlight() {
-    if (this.overlayElement) {
-      this.overlayElement.style.display = 'none';
+    if (this.overlayElem) {
+      this.overlayElem.style.display = 'none';
     }
   }
 
-  /*
-   *  @method  updateHighlightElement
-   *
-   *  @desc  Create an overlay element and set its position on the page.
-   *
-   *  @param  {Object}  element          -  DOM element node to highlight
-   *  @param  {String}  info             -  Description of the element
-   *
-   */
-
-   updateHighlightElement (element, info) {
-
-    let rect  = element.getBoundingClientRect();
-
-    let isHidden = false;
-
-    const rectLeft  = rect.left > offset ?
-                    Math.round(rect.left - offset + window.scrollX) :
-                    Math.round(rect.left + window.scrollX);
-
-    let left = rectLeft;
-
-    const rectWidth  = rect.left > offset ?
-                    Math.max(rect.width  + offset * 2, minWidth) :
-                    Math.max(rect.width, minWidth);
-
-    let width = rectWidth;
-
-    const rectTop    = rect.top > offset ?
-                    Math.round(rect.top  - offset + window.scrollY) :
-                    Math.round(rect.top + window.scrollY);
-
-    let top = rectTop;
-
-    const rectHeight   = rect.top > offset ?
-                    Math.max(rect.height + offset * 2, minHeight) :
-                    Math.max(rect.height, minHeight);
-
-    let height = rectHeight;
-
-    if ((rect.height < 3) || (rect.width < 3)) {
-      isHidden = true;
-    }
-
-    if ((rectTop < 0) || (rectLeft < 0)) {
-      isHidden = true;
-      if (element.parentNode) {
-        const parentRect = element.parentNode.getBoundingClientRect();
-
-        if ((parentRect.top > 0) && (parentRect.left > 0)) {
-          top = parentRect.top > offset ?
-                    Math.round(parentRect.top  - offset + window.scrollY) :
-                    Math.round(parentRect.top + window.scrollY);
-          left = parentRect.left > offset ?
-                    Math.round(parentRect.left - offset + window.scrollX) :
-                    Math.round(parentRect.left + window.scrollX);
-        }
-        else {
-          left = offset;
-          top = offset;
-        }
-      }
-      else {
-        left = offset;
-        top = offset;
-      }
-    }
-
-    this.overlayElem.style.left   = left   + 'px';
-    this.overlayElem.style.top    = top    + 'px';
-
-    if (isHidden) {
-      this.borderElem.textContent = 'Heading is hidden';
-      this.borderElem.classList.add('skip-to-hidden');
-      this.overlayElem.style.width  = 'auto';
-      this.overlayElem.style.height = 'auto';
-      this.borderElem.style.width  = 'auto';
-      this.borderElem.style.height = 'auto';
-      height = this.borderElem.getBoundingClientRect().height;
-      width  = this.borderElem.getBoundingClientRect().width;
-      if (rect.top > offset) {
-        height += offset + 2;
-        width += offset + 2;
-      }
-    }
-    else {
-      this.borderElem.textContent = '';
-      this.borderElem.classList.remove('skip-to-hidden');
-      this.overlayElem.style.width  = width  + 'px';
-      this.overlayElem.style.height = height + 'px';
-      this.borderElem.style.width  = (width  - 2 * borderWidth) + 'px';
-      this.borderElem.style.height = (height - 2 * borderWidth) + 'px';
-    }
-
-    this.overlayElem.style.display = 'block';
-
-    if (info) {
-      this.infoElem.style.display = 'inline-block';
-      this.infoElem.textContent = info;
-      if (top >= this.infoElem.getBoundingClientRect().height) {
-        this.borderElem.classList.remove('hasInfoBottom');
-        this.infoElem.classList.remove('hasInfoBottom');
-        this.borderElem.classList.add('hasInfoTop');
-        this.infoElem.classList.add('hasInfoTop');
-        if (!isHidden) {
-          this.infoElem.style.top = (-1 * (height + this.infoElem.getBoundingClientRect().height - 2 * borderWidth)) + 'px';
-        }
-        else {
-          this.infoElem.style.top = (-1 * (this.infoElem.getBoundingClientRect().height + this.borderElem.getBoundingClientRect().height)) + 'px';
-        }
-      }
-      else {
-        this.borderElem.classList.remove('hasInfoTop');
-        this.infoElem.classList.remove('hasInfoTop');
-        this.borderElem.classList.add('hasInfoBottom');
-        this.infoElem.classList.add('hasInfoBottom');
-        this.infoElem.style.top = -2 + 'px';
-      }
-      return this.infoElem;
-    }
-    else {
-      this.borderElem.classList.remove('hasInfo');
-      this.infoElem.style.display = 'none';
-      return this.overlayElem;
-    }
-  }
 }